requirements.txt
aiogram == 3.0.0b6
python-dotenv
dataclasses-json